requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
cryptography
imagehash
opencv-python-headless
pillow
